\name{exampleEigen-package}
\alias{exampleEigen-package}
\alias{exampleEigen}
\docType{package}
\title{
\packageTitle{exampleEigen}
}
\description{
\packageDescription{exampleEigen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{exampleEigen}
\packageIndices{exampleEigen}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{exampleEigen}

Maintainer: \packageMaintainer{exampleEigen}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
