\name{exampleFLGP-package}
\alias{exampleFLGP-package}
\alias{exampleFLGP}
\docType{package}
\title{
\packageTitle{exampleFLGP}
}
\description{
\packageDescription{exampleFLGP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{exampleFLGP}
\packageIndices{exampleFLGP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{exampleFLGP}

Maintainer: \packageMaintainer{exampleFLGP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
